requirements.txt
asgiref==3.5.2
asttokens==2.0.8
atomicwrites==1.4.1
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
dj-database-url==1.0.0
Django==4.1
django-ckeditor==6.5.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-js-asset==2.0.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.1.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
ipykernel==6.16.0
ipython==8.5.0
jdcal==1.4.1
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
oauth2==1.9.0.post1
openpyxl==3.0.10
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.1
Pygments==2.13.0
pyimgur==0.6.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pywin32==304
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.4.0
tzdata==2022.2
urllib3==1.26.12
wcwidth==0.2.5
whitenoise==6.2.0
gunicorn==20.0.4
